Add files via upload
The abalone data set that we are analyzing is a data set with various physical measurements of sea snails. In this notebook we'll be constructing linear models to map from input variables to output variables, and see how well we are able to predict the number of rings in abalone shells.